+++ trunk/mpf-rest-api/src/main/java/org/mitre/mpf/rest/api/node/EnvironmentVariableModel.java
@@ -26,22 +26,22 @@
 
 package org.mitre.mpf.rest.api.node;
 
-import org.mitre.mpf.nms.xml.EnvironmentVariable;
+import org.mitre.mpf.nms.json.EnvironmentVariable;
 
 public class EnvironmentVariableModel {
 	private String name; //env var vname
 	private String value; //env var value
 	private String sep;
-	
+
 	public EnvironmentVariableModel() { }
-	
+
 	public EnvironmentVariableModel(String name, String value, String sep) {
 		this.name = name;
 		this.value = value;
 		this.sep = sep;
 	}
-	
-	public EnvironmentVariableModel(EnvironmentVariable envVar) {	
+
+	public EnvironmentVariableModel(EnvironmentVariable envVar) {
 		this.name = envVar.getKey();
 		this.value = envVar.getValue();
 		this.sep = envVar.getSep();
@@ -50,7 +50,7 @@ public class EnvironmentVariableModel {
 	public String getName() {
 		return name;
 	}
-	
+
 	public String getValue() {
 		return value;
 	}
