+++ trunk/node-manager/src/main/java/org/mitre/mpf/nms/ServiceDescriptor.java
@@ -27,7 +27,7 @@
 package org.mitre.mpf.nms;
 
 import org.mitre.mpf.nms.NodeManagerConstants.ServiceTypes;
-import org.mitre.mpf.nms.xml.Service;
+import org.mitre.mpf.nms.json.Service;
 import org.springframework.stereotype.Component;
 
 import java.io.Serializable;
